Backend(Django)/requirements.txt
Django==5.1.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dj-database-url==1.2.0
psycopg2-binary==2.9.8
gunicorn==20.1.0
whitenoise==6.5.0
python-decouple==3.8
django-cors-headers==3.14.0